Allow Service Account FireCloud users to run FISS
I was trying to use FISS to call FireCloud using a service account that I had registered for use in FireCloud.  I realized I was not able to get through the proxy because the service account was not requesting the right scopes.  Added the correct userinfo.profile and userinfo.email scopes that are required for a service account to get through the proxies.  This has no negative effect upon a regular user of FISS.